requirements.txt
flask
beautifulsoup4
racing-report-korotova==0.0.14
pytest
pandas